temp fixed my-list page & ratings page done
Previously, my-list page will reload on movie/tv status update and the placeholder will show due to the usemutation data become undefined. Temp solution is to preserve the data using another state hook and update as soon as the usemutation data updates.

Better solutions is to only refetch that particular movie/tv data when updated and not the whole list.